backend/requirements.txt
fastapi
uvicorn
chromadb
openai
python-multipart
python-dotenv
